requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/base.txt requirements/base.in
#
alembic==1.4.3            # via flask-migrate
attrs==20.3.0             # via jsonschema, pytest
blinker==1.4              # via sentry-sdk
certifi==2020.6.20        # via requests, sentry-sdk
chardet==3.0.4            # via requests
click==7.1.2              # via flask
flasgger==0.9.5           # via -r requirements/base.in
flask-admin==1.5.7        # via -r requirements/base.in
flask-marshmallow==0.14.0  # via -r requirements/base.in
flask-migrate==2.5.3      # via -r requirements/base.in
flask-sqlalchemy==2.4.4   # via -r requirements/base.in, flask-migrate
flask==1.1.2              # via -r requirements/base.in, flasgger, flask-admin, flask-marshmallow, flask-migrate, flask-sqlalchemy, sentry-sdk
idna==2.10                # via requests
iniconfig==1.1.1          # via pytest
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
jsonschema==3.2.0         # via flasgger
mako==1.1.3               # via alembic
markupsafe==1.1.1         # via jinja2, mako, wtforms
marshmallow-enum==1.5.1   # via -r requirements/base.in
marshmallow-sqlalchemy==0.24.0  # via -r requirements/base.in
marshmallow==3.9.0        # via flask-marshmallow, marshmallow-enum, marshmallow-sqlalchemy
mistune==0.8.4            # via flasgger
packaging==20.4           # via pytest
pluggy==0.13.1            # via pytest
psycopg2-binary==2.8.6    # via -r requirements/base.in
py==1.9.0                 # via pytest
pydantic[dotenv]==1.6.1   # via -r requirements/base.in
pyjwt==1.7.1              # via -r requirements/base.in
pyparsing==2.4.7          # via packaging
pyrsistent==0.17.3        # via jsonschema
pytest==6.1.2             # via -r requirements/base.in
python-dateutil==2.8.1    # via alembic
python-dotenv==0.14.0     # via pydantic
python-editor==1.0.4      # via alembic
pyyaml==5.3.1             # via flasgger
requests==2.24.0          # via -r requirements/base.in
sentry-sdk[flask]==0.19.3  # via -r requirements/base.in
six==1.15.0               # via flasgger, flask-marshmallow, jsonschema, packaging, python-dateutil
sqlalchemy==1.3.19        # via alembic, flask-sqlalchemy, marshmallow-sqlalchemy
toml==0.10.2              # via pytest
urllib3==1.25.11          # via requests, sentry-sdk
werkzeug==1.0.1           # via flask
wtforms==2.3.3            # via flask-admin

# The following packages are considered to be unsafe in a requirements file:
# setuptools
